refactor: ListAdapter와 DiffUtil을 사용하여 RecyclerView 어댑터 개선

+++ app/src/main/java/woowacourse/shopping/feature/goodsdetails/GoodsDetailsActivity.kt
@@ -60,7 +60,7 @@ class GoodsDetailsActivity : AppCompatActivity() {
     private fun navigateToLastViewed() {
         viewModel.navigateToLastViewedCart.observe(this) { lastViewedCart ->
             lastViewedCart.let {
-                val intent = newIntent(this@GoodsDetailsActivity, it.id.toLong())
+                val intent = newIntent(this@GoodsDetailsActivity, it.id)
                 startActivity(intent)
                 finish()
             }
